Got rid of ImageRequestBuilder_v2 interface
